requirements.txt
# Core data science
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
imbalanced-learn==0.12.3

# NLP
spacy==3.7.4
transformers==4.41.2
torch==2.3.1

# API
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.3
httpx==0.27.0

# Experiment tracking & versioning
mlflow==2.13.2
dvc==3.51.2

# Interpretability
shap==0.45.1

# Auth
python-jose[cryptography]==3.3.0

# Logging
python-json-logger==2.0.7

# Config
pyyaml==6.0.1

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
hypothesis==6.103.1
